Support compiling the basic comparisons operations for the `F64` type.

+++ src/runtime.c
@@ -3,9 +3,9 @@
 #include <stdio.h>
 
 // this is the entry function to buraq
-extern double boot() asm("boot");
+extern int boot() asm("boot");
 
 int main() {
-    printf("%g\n", boot());
+    printf("%d\n", boot());
     return 0;
 }
